Criando excepitons personalizadas e Traduzindo codigo.

+++ src/main/java/com/example/collab/service/CollaboratorService.java
@@ -6,14 +6,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.example.collab.domain.model.Collaborator;
-import com.example.collab.domain.valueobject.bancario.Agencia;
-import com.example.collab.domain.valueobject.bancario.Banco;
-import com.example.collab.domain.valueobject.bancario.Conta;
-import com.example.collab.domain.valueobject.bancario.TipoConta;
-import com.example.collab.domain.valueobject.contato.Email;
-import com.example.collab.domain.valueobject.contato.Telefone;
+import com.example.collab.domain.valueobject.banking.Agency;
+import com.example.collab.domain.valueobject.banking.Bank;
+import com.example.collab.domain.valueobject.banking.Account;
+import com.example.collab.domain.valueobject.banking.TypeAccount;
+import com.example.collab.domain.valueobject.contact.Email;
+import com.example.collab.domain.valueobject.contact.Phone;
 import com.example.collab.dto.request.CollaboratorRequestDTO;
 import com.example.collab.dto.response.CollaboratorResponseDTO;
+import com.example.collab.exception.domain.InvalidCollaboratorException;
 import com.example.collab.mapper.CollaboratorMapper;
 import com.example.collab.repository.CollaboratorRepository;
 import com.example.collab.service.validation.CollaboratorValidator;
@@ -41,15 +42,21 @@ public class CollaboratorService {
                 req.getTituloEleitor());
 
         collaboratorValidator.validateNewCollaboratorBank(
-                req.getBanco(),
-                req.getConta(),
+                req.getAccount(),
                 req.getPix());
 
         Collaborator collaborator = collaboratorMapper.toEntity(req);
 
-        Collaborator savedCollaborator = collaboratorRepository.save(collaborator);
+        if (collaborator != null) {
 
-        return collaboratorMapper.toResponse(savedCollaborator);
+            Collaborator savedCollaborator = collaboratorRepository.save(collaborator);
+
+            return collaboratorMapper.toResponse(savedCollaborator);
+
+        }
+
+        throw new InvalidCollaboratorException("Error creating collaborator");
+
     }
 
     public List<CollaboratorResponseDTO> getAllCollaborators() {
@@ -69,6 +76,7 @@ public class CollaboratorService {
                 .orElseThrow(() -> new RuntimeException("Matricula não encontrada"));
 
         return collaboratorMapper.toResponse(collaborator);
+
     }
 
     public CollaboratorResponseDTO getCollaboratorByCPF(String CPF) {
@@ -77,14 +85,16 @@ public class CollaboratorService {
                 .orElseThrow(() -> new RuntimeException("CPF não encontrado"));
 
         return collaboratorMapper.toResponse(collaborator);
+
     }
 
-    public CollaboratorResponseDTO getCollaboratorByNome(String nome) {
+    public CollaboratorResponseDTO getCollaboratorByName(String name) {
 
-        Collaborator collaborator = collaboratorRepository.findByNome(nome)
+        Collaborator collaborator = collaboratorRepository.findByName(name)
                 .orElseThrow(() -> new RuntimeException("Nome não encontrado"));
 
         return collaboratorMapper.toResponse(collaborator);
+
     }
 
     public void deleteCollaboratorbyMatricula(Integer matricula) {
@@ -92,7 +102,14 @@ public class CollaboratorService {
         Collaborator collaborator = collaboratorRepository.findByMatricula(matricula)
                 .orElseThrow(() -> new RuntimeException("Colaborador não encontrado com a matrícula: " + matricula));
 
-        collaboratorRepository.delete(collaborator);
+        if (collaborator != null) {
+
+            collaboratorRepository.delete(collaborator);
+
+        }
+
+        throw new RuntimeException("Colaborador não encontrado com a matrícula: " + matricula);
+
     }
 
     public CollaboratorResponseDTO updateCollaborator(Integer matricula, CollaboratorRequestDTO req) {
@@ -100,19 +117,32 @@ public class CollaboratorService {
         Collaborator existingCollaborator = collaboratorRepository.findByMatricula(matricula)
                 .orElseThrow(() -> new RuntimeException("Colaborador não encontrado com a matrícula: " + matricula));
 
-        existingCollaborator.setNome(req.getNome());
-        existingCollaborator.setEstadoCivil(req.getEstadoCivil());
+        existingCollaborator.setName(req.getName());
+
+        existingCollaborator.setMaritalStatus(req.getMaritalStatus());
+
         existingCollaborator.setEmail(new Email(req.getEmail()));
-        existingCollaborator.setTelefone(new Telefone(req.getTelefone()));
+
+        existingCollaborator.setPhone(new Phone(req.getPhone()));
+
         existingCollaborator.setEndereco(req.getEndereco());
-        existingCollaborator.setTipoConta(new TipoConta(req.getTipoConta()));
-        existingCollaborator.setBanco(new Banco(req.getBanco()));
-        existingCollaborator.setAgencia(new Agencia(req.getAgencia()));
-        existingCollaborator.setConta(new Conta(req.getConta()));
+
+        existingCollaborator.setTypeAccount(new TypeAccount(req.getTypeAccount()));
+
+        existingCollaborator.setBank(new Bank(req.getBank()));
+
+        existingCollaborator.setAgency(new Agency(req.getAgency()));
+
+        existingCollaborator.setAccount(new Account(req.getAccount()));
+
         existingCollaborator.setContatoEmergencia(req.getContatoEmergencia());
-        existingCollaborator.setTelefoneEmergencia(new Telefone(req.getTelefoneEmergencia()));
+
+        existingCollaborator.setPhoneEmergency(new Phone(req.getPhoneEmergency()));
+
         existingCollaborator.setEscolaridade(req.getEscolaridade());
+
         existingCollaborator.setCurso(req.getCurso());
+
         existingCollaborator.setObservacoes(req.getObservacoes());
 
         Collaborator updatedCollaborator = collaboratorRepository.save(existingCollaborator);
